+++ exo.jcr.component.core/src/main/java/org/exoplatform/services/jcr/impl/storage/jdbc/optimisation/db/MultiDbJDBCConnection.java
@@ -1043,13 +1043,11 @@ public class MultiDbJDBCConnection extends CQJDBCStorageConnection
       {
          removeValuesStatement =
             dbConnection
-               .prepareStatement("DELETE FROM JCR_MVALUE WHERE PROPERTY_ID IN (SELECT ID FROM JCR_MITEM WHERE NAME = ? OR NAME = ?)");
-         removeValuesStatement.setString(1, Constants.JCR_LOCKISDEEP.getAsString());
-         removeValuesStatement.setString(2, Constants.JCR_LOCKOWNER.getAsString());
+               .prepareStatement("DELETE FROM JCR_MVALUE WHERE PROPERTY_ID IN (SELECT ID FROM JCR_MITEM WHERE NAME = '[http://www.jcp.org/jcr/1.0]lockIsDeep' OR NAME = '[http://www.jcp.org/jcr/1.0]lockOwner')");
 
-         removeItemsStatement = dbConnection.prepareStatement("DELETE FROM JCR_MITEM WHERE NAME = ? OR NAME = ?");
-         removeItemsStatement.setString(1, Constants.JCR_LOCKISDEEP.getAsString());
-         removeItemsStatement.setString(2, Constants.JCR_LOCKOWNER.getAsString());
+         removeItemsStatement =
+            dbConnection
+               .prepareStatement("DELETE FROM JCR_MITEM WHERE NAME = '[http://www.jcp.org/jcr/1.0]lockIsDeep' OR NAME = '[http://www.jcp.org/jcr/1.0]lockOwner'");
 
          removeValuesStatement.executeUpdate();
          removeItemsStatement.executeUpdate();
